+++ addOns/pscanrules/src/test/java/org/zaproxy/zap/extension/pscanrules/InformationDisclosureSuspiciousCommentsScanRuleUnitTest.java
@@ -26,15 +26,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import org.apache.commons.httpclient.URI;
 import org.apache.commons.httpclient.URIException;
 import org.junit.jupiter.api.Test;
-import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.network.HttpMalformedHeaderException;
 import org.parosproxy.paros.network.HttpMessage;
@@ -47,23 +43,8 @@ class InformationDisclosureSuspiciousCommentsScanRuleUnitTest
 
     @Override
     protected InformationDisclosureSuspiciousCommentsScanRule createScanner() {
+        InformationDisclosureSuspiciousCommentsScanRule.setPayloadProvider(null);
         return new InformationDisclosureSuspiciousCommentsScanRule();
-    }
-
-    @Override
-    public void setUpZap() throws Exception {
-        super.setUpZap();
-
-        Path xmlDir =
-                Files.createDirectories(
-                        Paths.get(
-                                Constant.getZapHome(),
-                                InformationDisclosureSuspiciousCommentsScanRule
-                                        .suspiciousCommentsListDir));
-        Path testFile =
-                xmlDir.resolve(
-                        InformationDisclosureSuspiciousCommentsScanRule.suspiciousCommentsListFile);
-        Files.write(testFile, Arrays.asList("# FixMeNot", "  FixMe  ", "TODO", "\t "));
     }
 
     protected HttpMessage createHttpMessageWithRespBody(String responseBody, String contentType)
@@ -131,7 +112,7 @@ class InformationDisclosureSuspiciousCommentsScanRuleUnitTest
         assertEquals(Alert.CONFIDENCE_LOW, alertsRaised.get(0).getConfidence());
         assertEquals("FIXME", alertsRaised.get(0).getEvidence());
         assertEquals(
-                wrapEvidenceOtherInfo("\\bFixMe\\b", line1, 1), alertsRaised.get(0).getOtherInfo());
+                wrapEvidenceOtherInfo("\\bFIXME\\b", line1, 1), alertsRaised.get(0).getOtherInfo());
     }
 
     @Test
@@ -175,7 +156,7 @@ class InformationDisclosureSuspiciousCommentsScanRuleUnitTest
         assertEquals("FIXME", alertsRaised.get(0).getEvidence());
         // detected 2 times, the first in the element
         assertEquals(
-                wrapEvidenceOtherInfo("\\bFixMe\\b", line1, 2), alertsRaised.get(0).getOtherInfo());
+                wrapEvidenceOtherInfo("\\bFIXME\\b", line1, 2), alertsRaised.get(0).getOtherInfo());
     }
 
     @Test
@@ -272,6 +253,47 @@ class InformationDisclosureSuspiciousCommentsScanRuleUnitTest
         assertFalse(ResourceIdentificationUtils.isJavaScript(msg));
 
         // When
+        scanHttpResponseReceive(msg);
+
+        // Then
+        assertEquals(0, alertsRaised.size());
+    }
+
+    @Test
+    void shouldAlertOnSuspiciousCommentProvidedByCustomPayload()
+            throws HttpMalformedHeaderException, URIException {
+
+        // Given
+        Iterable<String> customPayloads = List.of("zap_internal", "my_insights");
+        String body =
+                "<h1>Some text <!--MY_INSIGHTS: This is a test --></h1>\n"
+                        + "<b>Welcome to Zaproxy</b>\n";
+        HttpMessage msg = createHttpMessageWithRespBody(body, "text/html;charset=ISO-8859-1");
+
+        // When
+        InformationDisclosureSuspiciousCommentsScanRule.setPayloadProvider(() -> customPayloads);
+        scanHttpResponseReceive(msg);
+
+        // Then
+        assertEquals(1, alertsRaised.size());
+    }
+
+    @Test
+    void shouldNotAlertIfNeitherCustomNorStandardPayloadsFound()
+            throws HttpMalformedHeaderException, URIException {
+
+        // Given
+        Iterable<String> customPayloads = List.of("zap_internal", "my_insights");
+        String body =
+                "<h1>Some text <!-- Nothing special here --></h1>\n"
+                        + "<b>Welcome to Zaproxy</b>\n";
+        HttpMessage msg = createHttpMessageWithRespBody(body, "text/html;charset=ISO-8859-1");
+
+        assertTrue(msg.getResponseHeader().isText());
+        assertFalse(ResourceIdentificationUtils.isJavaScript(msg));
+
+        // When
+        InformationDisclosureSuspiciousCommentsScanRule.setPayloadProvider(() -> customPayloads);
         scanHttpResponseReceive(msg);
 
         // Then