+++ addOns/pscanrules/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Added 'DEBUG' to list of suspicious comments.
+- Added custom payload support (via Custom Payloads add-on) to 'Information Disclosure Suspicious Comments' scan rule.
+- Removed suspicious-comments.txt file in favor of payload editing via Custom Payloads add-on.
 
 ## [48] - 2023-05-03
 ### Added